+++ packages/patrol_cli/lib/src/commands/test.dart
@@ -13,7 +13,6 @@ import 'package:patrol_cli/src/coverage/web_coverage_tool.dart';
 import 'package:patrol_cli/src/crossplatform/app_options.dart';
 import 'package:patrol_cli/src/crossplatform/video_recording_config.dart';
 import 'package:patrol_cli/src/dart_defines_reader.dart';
-import 'package:patrol_cli/src/dashboard/dashboard_config.dart';
 import 'package:patrol_cli/src/devices.dart';
 import 'package:patrol_cli/src/ios/ios_test_backend.dart';
 import 'package:patrol_cli/src/macos/macos_test_backend.dart';
@@ -547,14 +546,7 @@ See https://github.com/leancodepl/patrol/issues/1316 to learn more.
       bitRate: int.tryParse(stringArg('video-bit-rate') ?? ''),
     );
 
-    // Passing a path is enough to ask for the report.
-    final htmlReportPath = stringArg('html-report-path');
-    final dashboardConfig = DashboardConfig(
-      enabled: boolArg('html-report') || htmlReportPath != null,
-      outputPath:
-          htmlReportPath ?? DashboardConfig.defaultOutputPath(testDirectory),
-      testDirectory: testDirectory,
-    );
+    final dashboardConfig = htmlReportConfig(testDirectory);
 
     switch (device.targetPlatform) {
       case TargetPlatform.android:
@@ -601,6 +593,7 @@ See https://github.com/leancodepl/patrol/issues/1316 to learn more.
           showFlutterLogs: showFlutterLogs,
           hideTestSteps: hideTestSteps,
           clearTestSteps: clearTestSteps,
+          dashboardConfig: dashboardConfig,
         );
     }