+++ packages/patrol_cli/lib/src/runner/patrol_command_runner.dart
@@ -166,6 +166,7 @@ class PatrolCommandRunner extends CompletionCommandRunner<int> {
     final webTestBackend = WebTestBackend(
       processManager: _processManager,
       parentDisposeScope: _disposeScope,
+      rootDirectory: rootDirectory,
       logger: _logger,
     );